requirements.txt
reportlab==4.2.0
pdfminer.six==20231228
python-docx==1.1.2
playwright==1.47.0
